+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## [v1.1.0]
+### Added
+- Added new setting `ttt_deadringer_corpse_role`
+    - Determines if a player's real role or a fake innocent role will be shown on the corpse
+- Added new setting `ttt_deadringer_corpse_confirm`
+    - Determines if the fake corpse will confirm the player's role when checked
+    - This may be used in tandem with `ttt_deadringer_corpse_role` to confirm fake/real roles
+
 ## [v1.0.1]
 ### Fixed
 - Fixed misleading text for the `ttt_deadringer_damage_reduction_time` setting to mention that the value is in absolute seconds and not a percentage.